updated stylesheet

+++ server/routes/private-data.js
@@ -141,8 +141,8 @@ router.delete('/volunteerRemove', function (req, res) {
       console.log(err);
       res.sendStatus(500);
     } else {
-      client.query('DELETE FROM role_user WHERE user_id=$1;',
-        [removeEntry.user_id], function (err, result) {
+      client.query('DELETE FROM role_user WHERE user_id=$1 AND role_id=$2;',
+        [removeEntry.user_id,removeEntry.role_id], function (err, result) {
 
           if (err) {
             console.log(err);
